Use ZMS in place of NVS

+++ lib/astarte_device_sdk/include/key_value/entry_hash.h
@@ -21,6 +21,6 @@
  * @param[in] key Used for generation of the hash.
  * @return The generated hash.
  */
-uint16_t astarte_key_value_entry_hash_generate(const char *namespace, const char *key);
+uint32_t astarte_key_value_entry_hash_generate(const char *namespace, const char *key);
 
 #endif // KEY_VALUE_ENTRY_HASH_H
